Move generate EV combinations to pokemon_helpers
