+++ agent/basic_pokemon_agent.py
@@ -5,6 +5,7 @@ from agent.base_agent import BaseAgent
 from pokemon_helpers.damage_stats import DamageStatCalc
 from pokemon_helpers.pokemon import Pokemon
 from pokemon_helpers.pokemon import calculate_spe_range
+from pokemon_helpers.pokemon import generate_all_ev_combinations
 
 from config import POKEMON_DATA
 
@@ -417,48 +418,3 @@ def calc_opp_position_helper(opp_gs):
             opp_posn += poke["pct_hp"]
 
     return opp_posn
-
-
-def generate_all_ev_combinations():
-    """Generate all possible stat investment combinations."""
-    combinations = {}
-
-    combinations["atk"] = []
-    combinations["spa"] = []
-    atk_combinations = []
-    atk_combinations.append((False, False))
-    atk_combinations.append((True, False))
-    atk_combinations.append((False, True))
-    atk_combinations.append((True, True))
-    for combination in atk_combinations:
-        result_dict = {}
-        result_dict["max_evs"] = combination[0]
-        result_dict["positive_nature"] = combination[1]
-        combinations["atk"].append(result_dict)
-        combinations["spa"].append(result_dict)
-
-    combinations["hp"] = []
-    combinations["def"] = []
-    combinations["spd"] = []
-
-    def_combinations = []
-    def_combinations.append((False, False))
-    def_combinations.append((False, False))
-    def_combinations.append((True, False))
-    def_combinations.append((True, False))
-    def_combinations.append((False, True))
-    def_combinations.append((False, True))
-    def_combinations.append((True, True))
-    def_combinations.append((True, True))
-
-    for combination in def_combinations:
-        result_dict = {}
-        result_dict["max_evs"] = combination[0]
-        result_dict["positive_nature"] = combination[1]
-        combinations["def"].append(result_dict)
-        combinations["spd"].append(result_dict)
-
-    combinations["hp"].append({"max_evs": True})
-    combinations["hp"].append({"max_evs": False})
-
-    return combinations
